placement-guardian-bot/backend/requirements.txt
# Placement Guardian Bot - Backend

## Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-telegram-bot>=20.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-firestore>=2.11.0
google-cloud-pubsub>=2.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cryptography>=41.0.0
httpx>=0.25.0
python-multipart>=0.0.6
firebase-admin>=6.2.0
